+++ feature/forecast/src/main/kotlin/com/weather/feature/forecast/components/hourlydata/DailyStaticData.kt
@@ -1,0 +1,137 @@
+package com.weather.feature.forecast.components.hourlydata
+
+import com.weather.model.Daily
+import com.weather.model.DailyPreview
+
+
+val DailyStaticData = listOf(
+    Daily(
+        0,
+        0.0,
+        "Tomorrow",
+        0,
+        0.0,
+        0,
+        0,
+        0.0,
+        0,
+        0,
+        0,
+        0.0,
+        0.0,
+        0.0,
+        0,
+        "Cloudy",
+        "",
+        "",
+        0,
+        0.0,
+        0.0
+    ),
+    Daily(
+        0,
+        0.0,
+        "Friday",
+        0,
+        0.0,
+        0,
+        0,
+        0.0,
+        0,
+        0,
+        0,
+        0.0,
+        0.0,
+        0.0,
+        0,
+        "Rain",
+        "",
+        "",
+        0,
+        0.0,
+        0.0
+    ),
+    Daily(
+        0,
+        0.0,
+        "Saturday",
+        0,
+        0.0,
+        0,
+        0,
+        0.0,
+        0,
+        0,
+        0,
+        0.0,
+        0.0,
+        0.0,
+        0,
+        "Cloudy",
+        "",
+        "",
+        0,
+        0.0,
+        0.0
+    ),
+    Daily(
+        0,
+        0.0,
+        "Sunday",
+        0,
+        0.0,
+        0,
+        0,
+        0.0,
+        0,
+        0,
+        0,
+        0.0,
+        0.0,
+        0.0,
+        0,
+        "Sunny",
+        "",
+        "",
+        0,
+        0.0,
+        0.0
+    )
+)
+val DailyPreviewStaticData = listOf(
+    DailyPreview(
+        tempDay = "283",
+        tempNight = "275",
+        time = "Tomorrow",
+        icon = "",
+        condition = "Clouds"
+    ),
+    DailyPreview(
+        tempDay = "280",
+        tempNight = "274",
+        time = "Wed",
+        icon = "",
+        condition = "Snow"
+    ),
+    DailyPreview(
+        tempDay = "284",
+        tempNight = "276",
+        time = "Thur",
+        icon = "",
+        condition = "Clouds"
+    ),
+    DailyPreview(
+        tempDay = "278",
+        tempNight = "270",
+        time = "Friday",
+        icon = "",
+        condition = "Rain"
+    ),
+    DailyPreview(
+        tempDay = "281",
+        tempNight = "269",
+        time = "Saturday",
+        icon = "",
+        condition = "Snow"
+    )
+)
